Finishing cleaning up the main file, particularly the main window.
There's still a lot of cleanup to do, but at least now it should be
more manageable and easier to work on. Has yet to be thoroughly tested,
but adding/removing subscriptions, and copying addresses seems to be
working. Withdrawing/sending manually has yet to be tested, nor has a
successful subscription payment been made. Some validation for
subscriptions has also been removed for now, but is intended to be
added in another form.