Adding test for paying subscription, refactoring a bit.
We needed to move some of this out of the infinite loop meant to be
stopped by the main thread when the UI is stopped. I don't know of any
way to replicate this in tests. Moved making the payment to the
subscription object, seemed more at home there since all of the
information needed was on the subscriptions class anyway.

This also moves the location of the monero_from_usd function to utils.