Adding wallet tests.
Not having an RPC Server running during the tests is a bit of a
problem when trying to test certain things, particularly creation of
a wallet. Seems impossible to test unless I figure out a way to get
the RPC server to run during the test. Could potentially use docker
for this.