Creating Views concept. A way to switch groups of interface elements.
All of the elements presented at once are grouped together and able to
be managed as a set. When switch_view is called, the elements for the
previous view will be destroyed and the new view elements will load.

There are other considerations that were needed to be made as well,
like the RPC Server observer functionality. This necessitated a change
in how we handle the server instances. Since they hinge on the wallet
being opened, I decided to allow for the potential to have multiple
wallets managed. We save each instance by the wallet name to the
RPCServer class _wallet_servers variable. This way we can access the
same instance across classes without having to pass around the object.
We retrieve the instance with a get('name') function.

In order to add a new View, you need to inherit from the View class,
and implement the build() function. For destruction of the View to
work properly, you need to add() each element you build.