Refactoring config variables as functions, removes wallet cli call.
We have had an issue with multiple processes trying to access the
wallet file that has caused flakiness that was difficult to figure
out. It turned out that we were trying to refresh the wallet through
the CLI, when it appears as though the wallet RPC handles this by
default.

This also allows for updating of the monero daemon that the RPC server
is connecting to and will restart the RPC process in order to connect
to the other server. We could potentially change this to using calls
against the RPC server to change the daemon instead.