Removing WALLET_ADDRESS variable from config file, using RPC server.
We might want to do something similar to what is done for the RPCServer
object where we save the instances in memory so we can access them
later. With this implementation it needs to create a new instance
frequently and has to wait for a response from the RPC server.