Adding more tests using VCR for the rpc client and subscriptions.
I am suspicious of the wallet test as it seemed to pass even without
starting a wallet rpc server and seems to take slightly longer than
others to run. Taking longer to run could be because the yaml file
is quite large considering the amount of data captured from getting
prices. I couldn't find anything actually wrong with it though so I'm
committing it. Should have an eye kept on it though.